requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pandas==2.1.3
aiohttp==3.9.1
python-dotenv==1.0.0
jinja2==3.1.2
plotly==5.18.0
certifi==2023.11.17
requests==2.31.0
schedule==1.2.1
matplotlib==3.8.2
pytest==7.4.3
python-multipart==0.0.6
pydantic==2.5.2 